tests all pass

+++ features/support/step_helpers.rb
@@ -45,7 +45,7 @@ end
 
 def log_in
   default_user
-  create_school 'Glenn Stephens'
+  create_school 'Glenn Stephens' unless @additional_student
   visit '/'
   fill_in 'Login', :with => @default_user.username
   fill_in 'Password', :with => @default_user.username
@@ -153,7 +153,7 @@ def default_user
   @default_user.roles = ['regular_user']
   @default_user.save!
 
-  SpecialUserGroup.delete_all
+  SpecialUserGroup.delete_all unless @additional_student
   # put other stuff above this
   @default_user
 end
